Add base controller, move contact to its own controller
